Revert "Refactor dependencies"
This reverts commit b47000a051d3c2f4687fbc3afb962c73a7484445.
